refactor(teleport): extract ensureRemoteProject into remote-project-ensure service
Move the inline ensureRemoteProject closure out of createLiveDeps() in
teleport-ops.ts into a standalone src/main/services/remote-project-ensure.ts
module so it can be reused by the upcoming launch-on-cloud RPC domain.
Behavior is unchanged: same clone-into-~/hive-projects/<name> semantics,
same reuse-by-origin-URL matching. TeleportOpsDeps/TeleportGit shapes are
untouched.
